% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{A small synthetic sample data set.}
\format{An object of class \code{list} of length 2.}
\usage{
example
}
\description{
A small data set containing generated segments to use with CoNVaQ.
}
\details{
A list containing two data frames.
\describe{
 \item{disease}{Data frame containing segments for the disease group}
 \item{healthy}{Data frame containing segments for the healthy group}
}
}
\keyword{datasets}
