% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{frequencies.convaq}
\alias{frequencies.convaq}
\title{Extract variation frequencies for each region as a table.}
\usage{
\method{frequencies}{convaq}(x)
}
\arguments{
\item{x}{A convaq object.}
}
\description{
Extract variation frequencies for each region as a table.
}
