% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{frequencies}
\alias{frequencies}
\title{Extract variation frequencies for each region as a data frame.}
\usage{
frequencies(x, ...)
}
\arguments{
\item{x}{A convaq object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Extract variation frequencies for each region as a data frame.
}
