% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.convaq}
\alias{print.convaq}
\title{Print description of CoNVaQ results object.}
\usage{
\method{print}{convaq}(x)
}
\arguments{
\item{x}{A convaq object.}
}
\description{
Print description of CoNVaQ results object.
}
