% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{regions.convaq}
\alias{regions.convaq}
\title{Extract matching CNV regions from CoNVaQ result as a table.}
\usage{
\method{regions}{convaq}(x)
}
\arguments{
\item{x}{A convaq object.}
}
\description{
Extract matching CNV regions from CoNVaQ result as a table.
}
