% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{regions}
\alias{regions}
\title{Extract matching CNV regions from CoNVaQ result as a table.}
\usage{
regions(x, ...)
}
\arguments{
\item{x}{A convaq object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Extract matching CNV regions from CoNVaQ result as a table.
}
