% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{states.convaq}
\alias{states.convaq}
\title{Extract states of individual samples from CoNVaQ result as a table.}
\usage{
\method{states}{convaq}(x)
}
\arguments{
\item{x}{A convaq object.}
}
\description{
Extract states of individual samples from CoNVaQ result as a table.
}
