% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{states}
\alias{states}
\title{Extract states of individual samples from CoNVaQ result as a table.}
\usage{
states(x, ...)
}
\arguments{
\item{x}{A convaq object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Extract states of individual samples from CoNVaQ result as a table.
}
