% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convaq.R
\name{convaq}
\alias{convaq}
\title{Perform CNV-based association study.}
\usage{
convaq(segments1, segments2, model, p.cutoff = 0.05, qvalues = FALSE,
  qvalues.rep = 500, nthreads = NULL)
}
\description{
Perform CNV-based association study.
}
